#110 Deactivate token validation

+++ server/controllers/likesController.ts
@@ -11,33 +11,40 @@ const likesController = {
   },
 
   addOrCancelLike: async (req: Request, res: Response) => {
+    const tokenValidity = true; // tokenAuthentication(req);
     const { user_id, post_id } = req.body;
     const likedOrNot = await Liking.findOne({ where: { user_id, post_id } });
 
-    try {
-      if (likedOrNot) {
-        await Liking.destroy({ where: { user_id, post_id } })
-          .then(() => {
-            likesController.updateLikesCount(post_id);
-          })
-          .then(() => {
-            res
-              .status(status.OK)
-              .json({ message: `Canceled the like of the post ${post_id}` });
-          });
-      } else {
-        await Liking.create({ user_id, post_id })
-          .then(() => {
-            likesController.updateLikesCount(post_id);
-          })
-          .then(() => {
-            res
-              .status(status.CREATED)
-              .json({ message: `Liked the post ${post_id}` });
-          });
+    if (!tokenValidity) {
+      res.status(status.UNAUTHORIZED).json({ message: 'Invalid Token' });
+    } else {
+      try {
+        if (likedOrNot) {
+          await Liking.destroy({ where: { user_id, post_id } })
+            .then(() => {
+              likesController.updateLikesCount(post_id);
+            })
+            .then(() => {
+              res
+                .status(status.OK)
+                .json({ message: `Canceled the like of the post ${post_id}` });
+            });
+        } else {
+          await Liking.create({ user_id, post_id })
+            .then(() => {
+              likesController.updateLikesCount(post_id);
+            })
+            .then(() => {
+              res
+                .status(status.CREATED)
+                .json({ message: `Liked the post ${post_id}` });
+            });
+        }
+      } catch (err) {
+        res
+          .status(status.NOT_FOUND)
+          .json({ message: 'Wrong id number requested' });
       }
-    } catch (err) {
-      res.status(status.NOT_FOUND).json({ message: 'Wrong id number requested' });
     }
   },
 };
